backend/libs/dnp_azure_auth/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "dnp-azure-auth"
version = "0.1.0"
dependencies = [
  "azure-identity",
  "azure-core",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
